% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_stream.R
\name{as_reader}
\alias{as_reader}
\title{Coerce an object to a readable stream}
\usage{
as_reader(x, queue_strategy)
}
\arguments{
\item{x}{object to coerce}

\item{queue_strategy}{optionally specify the queue strategy of the stream}
}
\value{
\code{streams::ReadableStream}
}
\description{
\code{as_reader} turns an existing object into a readable stream.
Coercion can be used when you want to create a readable stream from a known
data source. In general, this is the easiest way to create a readable stream.

As an S3 generic, \code{as_reader} holds methods for:
\itemize{
\item \code{base::data.frame}: creates a generator function that pushes
one row of a data.frame at a time into the stream
\item connection: pulls chunks from a connection line by line, closing when a 0-length
string is returned
\item \code{promises::promise}: uses the eventual output (\code{onFulfilled} value) of the promise as a chunk source.
If a rejection occurs, the stream will be errored.
\item default: create an iterator through \code{coro::as_iterator},
which is then used to supply chunks to the stream
}
}
\examples{
# create a readable stream from a numeric vector
reader <- as_reader(1:10)
on(reader, "data", print)

}
