% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_stream.R
\name{as_writer}
\alias{as_writer}
\title{Coerce an object to a writeable stream}
\usage{
as_writer(x, queue_strategy)
}
\arguments{
\item{x}{object to coerce}

\item{queue_strategy}{optionally specify the queue strategy of the stream}
}
\value{
\code{streams::WriteableStream}
}
\description{
\code{as_writer} turns an existing object into a writeable stream.
Coercion can be used when you want to write to a known
data sink In general, this is the easiest way to create a writeable stream.

As an S3 generic, \code{as_writer} holds methods for:
\itemize{
\item character: creates a stream that appends all chunks to a variable specified by the character input
\item connection: write to a file, coercing chunks to a character vector where required
\item default: throws an error
}
}
\examples{
# create a writeable stream from a file path
writer <- as_writer(file(tempfile()))

}
