% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{batch_stream}
\alias{batch_stream}
\title{Batch stream}
\usage{
batch_stream(size, ...)
}
\arguments{
\item{size}{the maximum length of each batch}

\item{...}{arguments passed to TransformStream}
}
\description{
Chunks that flow through a stream will be grouped
into lists of length \code{size} before being enqueued
}
\examples{
as_reader(1:100) \%|>\%
batch_stream(10) |>
   on("data", print)
}
\seealso{
Other functionals: 
\code{\link{collect_stream}()},
\code{\link{filter_stream}()},
\code{\link{flatten_stream}()},
\code{\link{limit_stream}()},
\code{\link{map_stream}()},
\code{\link{reduce_stream}()},
\code{\link{walk_stream}()}
}
\concept{functionals}
