% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{collect_stream}
\alias{collect_stream}
\title{Collect stream}
\usage{
collect_stream(...)
}
\arguments{
\item{...}{arguments passed to reduce_stream}
}
\description{
Form of reduce stream where all data is appended to a list before
being emitted.
}
\examples{
as_reader(1:10) \%|>\%
    collect_stream() |>
    on("data", print)

}
\seealso{
Other functionals: 
\code{\link{batch_stream}()},
\code{\link{filter_stream}()},
\code{\link{flatten_stream}()},
\code{\link{limit_stream}()},
\code{\link{map_stream}()},
\code{\link{reduce_stream}()},
\code{\link{walk_stream}()}
}
\concept{functionals}
