% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_emitter.R
\name{EventEmitter}
\alias{EventEmitter}
\title{Event Emitter}
\description{
Event Emitter

Event Emitter
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventEmitter-on}{\code{EventEmitter$on()}}
\item \href{#method-EventEmitter-prepend_listener}{\code{EventEmitter$prepend_listener()}}
\item \href{#method-EventEmitter-once}{\code{EventEmitter$once()}}
\item \href{#method-EventEmitter-prepend_once_listener}{\code{EventEmitter$prepend_once_listener()}}
\item \href{#method-EventEmitter-off}{\code{EventEmitter$off()}}
\item \href{#method-EventEmitter-remove_all_listeners}{\code{EventEmitter$remove_all_listeners()}}
\item \href{#method-EventEmitter-emit}{\code{EventEmitter$emit()}}
\item \href{#method-EventEmitter-listener_count}{\code{EventEmitter$listener_count()}}
\item \href{#method-EventEmitter-raw_listeners}{\code{EventEmitter$raw_listeners()}}
\item \href{#method-EventEmitter-event_names}{\code{EventEmitter$event_names()}}
\item \href{#method-EventEmitter-set_max_listeners}{\code{EventEmitter$set_max_listeners()}}
\item \href{#method-EventEmitter-get_max_listeners}{\code{EventEmitter$get_max_listeners()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-on"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-on}{}}}
\subsection{Method \code{on()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$on(event_name, fn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_name}}{TODOC}

\item{\code{fn}}{TODOC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-prepend_listener"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-prepend_listener}{}}}
\subsection{Method \code{prepend_listener()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$prepend_listener(event_name, fn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_name}}{TODOC}

\item{\code{fn}}{TODOC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-once"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-once}{}}}
\subsection{Method \code{once()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$once(event_name, fn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_name}}{TODOC}

\item{\code{fn}}{TODOC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-prepend_once_listener"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-prepend_once_listener}{}}}
\subsection{Method \code{prepend_once_listener()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$prepend_once_listener(event_name, fn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_name}}{TODOC}

\item{\code{fn}}{TODOC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-off"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-off}{}}}
\subsection{Method \code{off()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$off(event_name, fn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_name}}{TODOC}

\item{\code{fn}}{TODOC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-remove_all_listeners"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-remove_all_listeners}{}}}
\subsection{Method \code{remove_all_listeners()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$remove_all_listeners(event_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_name}}{TODOC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-emit"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-emit}{}}}
\subsection{Method \code{emit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$emit(event_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_name}}{TODOC}

\item{\code{...}}{arguments to pass to functions registered to a given event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-listener_count"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-listener_count}{}}}
\subsection{Method \code{listener_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$listener_count(event_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_name}}{TODOC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-raw_listeners"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-raw_listeners}{}}}
\subsection{Method \code{raw_listeners()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$raw_listeners(event_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_name}}{TODOC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-event_names"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-event_names}{}}}
\subsection{Method \code{event_names()}}{
TODOC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$event_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-set_max_listeners"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-set_max_listeners}{}}}
\subsection{Method \code{set_max_listeners()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$set_max_listeners(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{TODOC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventEmitter-get_max_listeners"></a>}}
\if{latex}{\out{\hypertarget{method-EventEmitter-get_max_listeners}{}}}
\subsection{Method \code{get_max_listeners()}}{
TODOC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$get_max_listeners()}\if{html}{\out{</div>}}
}

}
}
