% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{filter_stream}
\alias{filter_stream}
\title{Filter stream}
\usage{
filter_stream(fn, ...)
}
\arguments{
\item{fn}{function applied to each chunk of stream that returns a boolean value}

\item{...}{arguments passed to TransformStream}
}
\description{
Filter a stream based on a boolean function, only passing data that results in a
true condition
}
\examples{
filterer <- as_reader(1:10) \%|>\%
    filter_stream(~ .x < 5L) |>
    on("data", print)

}
\seealso{
Other functionals: 
\code{\link{batch_stream}()},
\code{\link{collect_stream}()},
\code{\link{flatten_stream}()},
\code{\link{limit_stream}()},
\code{\link{map_stream}()},
\code{\link{reduce_stream}()},
\code{\link{walk_stream}()}
}
\concept{functionals}
