% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{flatten_stream}
\alias{flatten_stream}
\title{Flatten stream}
\usage{
flatten_stream(...)
}
\arguments{
\item{...}{arguments passed to TransformStream}
}
\description{
Chunks that pass through the stream are flattened one-level and their
individual elements are enqueued
}
\seealso{
Other functionals: 
\code{\link{batch_stream}()},
\code{\link{collect_stream}()},
\code{\link{filter_stream}()},
\code{\link{limit_stream}()},
\code{\link{map_stream}()},
\code{\link{reduce_stream}()},
\code{\link{walk_stream}()}
}
\concept{functionals}
