% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_helpers.R
\name{is_readable_stream}
\alias{is_readable_stream}
\title{TODOC}
\usage{
is_readable_stream(stream)
}
\description{
TODOC
}
\seealso{
Other classes: 
\code{\link{is_transform_stream}()},
\code{\link{is_writeable_stream}()}
}
\concept{classes}
