% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_helpers.R
\name{is_writeable_stream}
\alias{is_writeable_stream}
\title{TODOC}
\usage{
is_writeable_stream(stream)
}
\description{
TODOC
}
\seealso{
Other classes: 
\code{\link{is_readable_stream}()},
\code{\link{is_transform_stream}()}
}
\concept{classes}
