% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{limit_stream}
\alias{limit_stream}
\title{Limit stream}
\usage{
limit_stream(max_length = 0L, size_fn, ...)
}
\arguments{
\item{max_length}{the maximum number of chunks that can run through a stream}

\item{size_fn}{a function that is used to increment the size counter}

\item{...}{arguments passed to TransformStream}
}
\description{
Limit the number of chunks that can run through a stream, closing
once a maximum length is reached
}
\examples{
limiter <- as_reader(1:10) \%|>\%
    limit_stream(5L, length) |>
    on("data", print)

}
\seealso{
Other functionals: 
\code{\link{batch_stream}()},
\code{\link{collect_stream}()},
\code{\link{filter_stream}()},
\code{\link{flatten_stream}()},
\code{\link{map_stream}()},
\code{\link{reduce_stream}()},
\code{\link{walk_stream}()}
}
\concept{functionals}
