% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{map_stream}
\alias{map_stream}
\title{Map stream}
\usage{
map_stream(fn, ...)
}
\arguments{
\item{fn}{function applied to each chunk of stream, the result of which is enqueued to
the readable side}

\item{...}{arguments passed to TransformStream}
}
\description{
Apply function \code{fn} to each chunk of data
}
\examples{
mapper <- as_reader(1:10) \%|>\%
    map_stream(as.character) |>
    on("data", print)

}
\seealso{
Other functionals: 
\code{\link{batch_stream}()},
\code{\link{collect_stream}()},
\code{\link{filter_stream}()},
\code{\link{flatten_stream}()},
\code{\link{limit_stream}()},
\code{\link{reduce_stream}()},
\code{\link{walk_stream}()}
}
\concept{functionals}
