% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{data}
\name{no_backpressure_strategy}
\alias{no_backpressure_strategy}
\title{No backpressure strategy}
\format{
An object of class \code{environment} of length 4.
}
\usage{
no_backpressure_strategy
}
\description{
A queue strategy where there is no highwater mark specified, meaning that the queue size is potentially infinite.
Specifically, the function \code{length} is used, with a highwater mark of \code{Inf}.
}
\seealso{
Other queue_strategies: 
\code{\link{QueueStrategy}},
\code{\link{byte_size_strategy}},
\code{\link{object_length_strategy}}
}
\concept{queue_strategies}
\keyword{datasets}
