% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{data}
\name{object_length_strategy}
\alias{object_length_strategy}
\title{Object length strategy}
\format{
An object of class \code{environment} of length 4.
}
\usage{
object_length_strategy
}
\description{
A queue strategy where the queue size is determined by the length of the object.
Specifically, the function \code{length} is used, with a highwater mark of \code{1L}.
}
\seealso{
Other queue_strategies: 
\code{\link{QueueStrategy}},
\code{\link{byte_size_strategy}},
\code{\link{no_backpressure_strategy}}
}
\concept{queue_strategies}
\keyword{datasets}
