% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-wrappers.R
\name{on}
\alias{on}
\title{On event}
\usage{
on(emitter, event, fn)
}
\arguments{
\item{emitter}{1}

\item{event}{1}

\item{fn}{1}
}
\description{
1
}
\seealso{
Other events: 
\code{\link{once}()}
}
\concept{events}
