% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-wrappers.R
\name{once}
\alias{once}
\title{Once event}
\usage{
once(emitter, event, fn)
}
\arguments{
\item{emitter}{1}

\item{event}{1}

\item{fn}{1}
}
\description{
1
}
\seealso{
Other events: 
\code{\link{on}()}
}
\concept{events}
