% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_emitter.R
\name{promise_once}
\alias{promise_once}
\title{Async once-event}
\usage{
promise_once(emitter, event)
}
\arguments{
\item{emitter}{an EventEmitter object}

\item{event}{character vector}
}
\description{
Async once-event
}
\concept{async}
