% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{QueueStrategy}
\alias{QueueStrategy}
\title{Queue Strategy}
\description{
1
}
\seealso{
Other queue_strategies: 
\code{\link{byte_size_strategy}},
\code{\link{no_backpressure_strategy}},
\code{\link{object_length_strategy}}
}
\concept{queue_strategies}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{highwater_mark}}{A number representing the queue size at which a stream will stop attempting to receive more chunks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{QueueStrategy$new()}}
\item \href{#method-NA-size}{\code{QueueStrategy$size()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
Create a new queue strategy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueueStrategy$new(highwater_mark, size_algorithm)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{highwater_mark}}{number denoting when a stream should stop asking for more data}

\item{\code{size_algorithm}}{function that is used to measure the size of the stream queue}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-size"></a>}}
\if{latex}{\out{\hypertarget{method-NA-size}{}}}
\subsection{Method \code{size()}}{
A function that is applied to a chunk in order to determine its size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueueStrategy$size(chunk)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chunk}}{a data chunk}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
numeric vector
}
}
}
