% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{reduce_stream}
\alias{reduce_stream}
\title{Reduce stream}
\usage{
reduce_stream(fn, base_value = 0L, ...)
}
\arguments{
\item{fn}{a function that is applied to the current accumulation and the current chunk}

\item{base_value}{the base value of the accumulation}

\item{...}{arguments passed to TransformStream}
}
\description{
Reduce stream data into one chunk via the iterative application of function \code{fn} to an accumulator and
each chunk

See also \code{base::Reduce}
}
\examples{
as_reader(1:100) \%|>\%
    reduce_stream(sum) |>
    on("data", print)
}
\seealso{
Other functionals: 
\code{\link{batch_stream}()},
\code{\link{collect_stream}()},
\code{\link{filter_stream}()},
\code{\link{flatten_stream}()},
\code{\link{limit_stream}()},
\code{\link{map_stream}()},
\code{\link{walk_stream}()}
}
\concept{functionals}
