% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_truncated_normal.R
\name{sample_truncated_normal}
\alias{sample_truncated_normal}
\title{Sample from a Truncated Normal Distribution}
\usage{
sample_truncated_normal(mean, sd, a, b)
}
\arguments{
\item{mean}{Numeric; mean of the normal distribution.}

\item{sd}{Numeric; standard deviation of the normal distribution.}

\item{a}{Numeric; lower bound of truncation.}

\item{b}{Numeric; upper bound of truncation.}
}
\value{
A numeric value sampled from the truncated normal distribution.
}
\description{
Draws a single random value from a truncated normal distribution with specified mean, standard deviation,
and lower/upper bounds.
}
\examples{
\dontrun{
sample_truncated_normal(mean = 10, sd = 2, a = 5, b = 15)
}

}
