% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-wrappers.R
\name{streampipe}
\alias{streampipe}
\alias{\%|>\%}
\title{Stream pipe}
\usage{
lhs \%|>\% rhs
}
\arguments{
\item{lhs}{tbd}

\item{rhs}{tbd}
}
\description{
Pipe the left-hand side stream into the right-hand side stream.
Returns the right-hand side for further use.
}
\seealso{
Other pipes: 
\code{\link{\%T>\%}()}
}
\concept{pipes}
