% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{summary.ReadableStream}
\alias{summary.ReadableStream}
\title{These are designed to be reminiscent of
the summary produced by connections.}
\usage{
\method{summary}{ReadableStream}(object, ...)
}
\arguments{
\item{object}{todo}

\item{...}{todo}
}
\description{
1
}
\details{
Basically, try to make streams feel like
the connection objects that people are familiar with0
ReadableStream summary
}
