% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{summary.TransformStream}
\alias{summary.TransformStream}
\title{TransformStream summary}
\usage{
\method{summary}{TransformStream}(object, ...)
}
\arguments{
\item{object}{todo}

\item{...}{todo}
}
\description{
1
}
