% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{summary.WriteableStream}
\alias{summary.WriteableStream}
\title{WriteableStream summary}
\usage{
\method{summary}{WriteableStream}(object, ...)
}
\arguments{
\item{object}{todo}

\item{...}{todo}
}
\description{
1
}
