% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-wrappers.R
\name{\%T>\%}
\alias{\%T>\%}
\title{Tee pipe}
\usage{
lhs \%T>\% rhs
}
\arguments{
\item{lhs}{stream to be teed}

\item{rhs}{stream to be piped to}
}
\description{
Tees a readablestream, piping the left branch
into the right-hand side stream, and returning
the right branch for further use.
}
\examples{
# create a stream that prints the data sent to it
# and also increments each chunk by one
stream <- as_reader(1:100) \%T>\%
    walk_stream(function(data) print(data)) \%|>\%
    map_stream(function(data) data + 1)

}
\seealso{
Other pipes: 
\code{\link{streampipe}}
}
\concept{pipes}
