% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_example.R
\docType{data}
\name{toy_example}
\alias{toy_example}
\title{Toy multi-state panel dataset}
\format{
A data frame with 7399 rows and 14 variables:
\describe{
\item{patient_id}{Unique identifier for each patient (numeric).}
\item{dead}{Binary indicator (0/1) for whether the patient is dead.}
\item{death_time}{Time of death if it has occurred or censoring time otherwise (numeric).}
\item{onset}{Binary indicator (0/1) for disease onset.}
\item{onset_age}{Age at disease onset if it has occurred or \code{death_time} otherwise (numeric).}
\item{age}{Patient's age at that specific visit (numeric).}
\item{visits}{Visit index (numeric/integer).}
\item{cov1}{Binary covariate.}
\item{cov2}{Binary covariate.}
\item{cov3}{Numeric covariate.}
}
The dataset may contain additional simulation-derived columns that are not
required by the main pipeline. The number of rows per patient corresponds to
the total number of visits he attended.
}
\source{
Simulated data assuming age as a time scale for an illness-death model
under a Markov assumption.
}
\description{
A simulated dataset used to illustrate the STREAMS workflow.
}
\examples{
data("toy_example")
str(toy_example)

}
\keyword{datasets}
