% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{walk_stream}
\alias{walk_stream}
\title{Walk stream}
\usage{
walk_stream(fn, ...)
}
\arguments{
\item{fn}{function applied to each chunk of the stream,
but does not enqueue the changed chunk}

\item{...}{arguments passed to TransformStream}
}
\description{
Map stream, but for side-effects
}
\examples{
as_reader(1:10) \%|>\%
 walk_stream(print) |>
 on("data", print)

}
\seealso{
Other functionals: 
\code{\link{batch_stream}()},
\code{\link{collect_stream}()},
\code{\link{filter_stream}()},
\code{\link{flatten_stream}()},
\code{\link{limit_stream}()},
\code{\link{map_stream}()},
\code{\link{reduce_stream}()}
}
\concept{functionals}
