% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit Semi-Parametric Multi-State Models for a Progressive Illness-Death Process}
\usage{
fit_model(
  data,
  cov_vector,
  clock_assumption,
  distribution,
  custom_formula = NULL
)
}
\arguments{
\item{data}{A data frame containing the original dataset in wide format.
It must include:
\itemize{
\item \code{patient_id}: unique individual identifier
\item \code{onset_age}: age at disease onset
\item \code{death_time}: age/time at death
\item \code{onset}: indicator of disease onset
\item \code{dead}: indicator of death
\item \code{age}: baseline age
\item all covariates listed in \code{cov_vector}
}}

\item{cov_vector}{A character vector of covariate names to be included
in the model.}

\item{clock_assumption}{A string specifying the time scale used for model fitting.
Accepted values are:
\itemize{
\item \code{"forward"}: all transitions use the forward timescale \code{Surv(Tstart, Tstop, status)}
\item \code{"mix"}: transitions 1 and 2 use forward time; transition 3 (disease → death)
uses a reset timescale \code{Surv(time, status)}
}}

\item{distribution}{A character string specifying the parametric distribution passed to
\code{\link[flexsurv]{flexsurvreg}} (e.g., \code{"weibull"}, \code{"gompertz"}, \code{"lognormal"}).}

\item{custom_formula}{An optional \code{Surv()} formula provided by the user.
If supplied, it replaces the automatically generated formulas and will be used
for \strong{all transitions}, regardless of the clock assumption.
This allows specification of nonlinear or interaction effects.
If \code{NULL} (default), the model formula is constructed automatically from \code{cov_vector}.}
}
\value{
A list of three fitted \code{flexsurvreg} model objects,
corresponding to each transition of the illness–death model.
}
\description{
This function fits parametric models to the three transitions of a progressive
illness–death model
using the \pkg{mstate} and \pkg{flexsurv} frameworks.
It supports both a \strong{forward timescale} (Markov process) and a \strong{mixed timescale} (Semi-Markov process)
through the argument \code{clock_assumption}.
}
\details{
The user may either:
\itemize{
\item provide a \strong{custom survival formula} (allowing splines, nonlinear effects, interactions, etc.),
\item or rely on the \strong{default formula} constructed automatically from the covariates in \code{cov_vector} assuming linear effects.
}
}
\examples{
\dontrun{

# Default model with linear covariate effects
fit_model(
  data = df,
  cov_vector = c("sex", "bmi"),
  clock_assumption = "forward",
  distribution = "weibull"
)

# Custom formula with spline effect on BMI and reset of timescale in the disease status
fit_model(
  data = df,
  cov_vector = c("sex", "bmi"),
  clock_assumption = "mix",
  distribution = "gompertz",
  custom_formula = survival::Surv(time, status) ~
      sex + splines::ns(bmi, df = 3)
)
}

}
